chore: organize store mutations + remove persistence layer
close #8
