--- packages/lib/src/server/persistence/index.ts
@@ -1,3 +1,0 @@
-export * from "./kv";
-export * from "./types";
-export * from "./convex";
