+++ packages/lib/src/server/stripe/misc.ts
@@ -1,7 +1,8 @@
-import Stripe from "stripe";
+import { Infer } from "convex/values";
 
 import { Implementation } from "../helpers";
 import { STRIPE_SUB_CACHE } from "../schema/tables";
+import { StoreInputValidator } from "../store";
 
 export const getSubscriptionImplementation: Implementation<
   {
@@ -9,21 +10,19 @@ export const getSubscriptionImplementation: Implementation<
   },
   Promise<STRIPE_SUB_CACHE>
 > = async (context, args, configuration) => {
-  const stripeCustomerId =
-    await configuration.persistence.getStripeCustomerIdByEntityId(
-      context,
-      args.entityId
-    );
+  const stripeCustomerId = await context.runMutation(configuration.store, {
+    type: "getStripeCustomerIdByEntityId",
+    entityId: args.entityId,
+  } as Infer<typeof StoreInputValidator>);
 
   if (!stripeCustomerId) {
     return { status: "none" } as STRIPE_SUB_CACHE;
   }
 
-  const data =
-    await configuration.persistence.getSubscriptionDataByStripeCustomerId(
-      context,
-      stripeCustomerId
-    );
+  const data = await context.runMutation(configuration.store, {
+    type: "getSubscriptionDataByStripeCustomerId",
+    stripeCustomerId: stripeCustomerId,
+  } as Infer<typeof StoreInputValidator>);
 
   if (!data) {
     return { status: "none" } as STRIPE_SUB_CACHE;
